+++ src/rent/reservation.js
@@ -394,12 +394,20 @@ require(['jquery', 'YSDRemoteDataSource','YSDMemoryDataSource','YSDSelectSelecto
                     controller.btnPaymentClick();
                     return false;
                 },
-
+                errorClass: 'text-danger',
                 rules : {
-                    'payment': 'required'
+                    'payment_method_id': 'required'
                 },
                 messages: {
-                    'payment': i18next.t('myReservation.paymentForm.paymentRequired')
+                    'payment_method_id': i18next.t('myReservation.pay.paymentMethodRequired')
+                },
+                errorPlacement : function(error, element) {
+                  if (element.attr('name') == 'payment_method_id')  {
+                     error.insertBefore('#btn_pay');
+                  }
+                  else {
+                     error.insertAfter(element);
+                  }
                 }
             }
         );